refactor: 移除 Yak Mcp 旧弹窗，全部入口统一跳转设置中心
